chore: move browser utils to new directory
