+++ src/browser/utils/deferNetworkRequestsUntil.ts
