+++ src/browser/utils/getAbsoluteWorkerUrl.ts
